sm8250-common: DeviceSettings: Update HBM tile & switch on screen off
What happens is that HBM is being turned off in lower level and the node is not being updated.
Writing 0 to node when the screen turns off to go around that.

Service only runs when HBM is turned on.
it stops itself when the screen turns off / HBM is turned off

Also remove some useless code all around and slightly improve...

Signed-off-by: chandu078 <chandudyavanapelli03@gmail.com>